requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
fastapi-mail==1.4.1
razorpay==1.4.1
pillow==10.2.0
reportlab==4.1.0
python-dotenv==1.0.0
pydantic[email]==2.5.3
pydantic-settings==2.2.1
httpx==0.27.0
slowapi==0.1.9
bcrypt==4.0.1
